test: add missing tests for quadratic_formula.py
