+++ Math/Algebra/Polynomials/factor_theorem.py
@@ -1,6 +1,7 @@
 # Factor theorem: check if (x - a) is a factor of a polynomial
 import math
 from typing import List, Union
+import math
 
 
 def evaluate_polynomial(coefficients: List[Union[int, float]], powers: List[Union[int, float]], x: Union[int, float]) -> float:
@@ -36,5 +37,7 @@ def check_factor(coefficients: List[Union[int, float]], powers: List[Union[int, 
     result = evaluate_polynomial(coefficients, powers, x)
     
     # Use math.isclose to account for floating-point precision issues
+    # Use math.isclose to handle floating-point precision issues
     # Use math.isclose to handle potential floating-point inaccuracies
     return math.isclose(result, 0.0, abs_tol=1e-9)
+    return math.isclose(result, 0, abs_tol=1e-9)
