+++ Math/Calculus/Integration/NumIntegration.py
@@ -18,8 +18,6 @@ def integrate_polynomial(coefficients: List[Union[int, float]], powers: List[Uni
     
     # Apply integration rule: ∫(ax^n)dx = (a/(n+1))×x^(n+1) + C
     for coeff, power in zip(coefficients, powers):
-        if power == -1:
-            raise ValueError("Integration of x^-1 results in ln|x|, which is not supported by this polynomial integration function.")
         new_power = power + 1
         new_coeff = coeff / new_power
         integrated_coeffs.append(new_coeff)
