+++ Math/Discrete_Math/Combinatorics/trinomial_theorem.py
@@ -3,6 +3,7 @@ import sys
 import os
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
 from combination import nCr
+from typing import List
 
 
 def trinomial_coefficient(n: int, i: int, j: int) -> int:
