+++ Math/Discrete_Math/Number_Theory/gcd.py
@@ -1,6 +1,5 @@
 # Greatest Common Divisor (GCD) calculation
 from typing import List
-from collections import Counter
 
 
 def prime_factorization_for_gcd(n: int) -> List[int]:
@@ -47,11 +46,8 @@ def compute_gcd(a: int, b: int) -> int:
     common_factors = set(factors_a) & set(factors_b)
     
     # Calculate GCD by taking minimum power of each common prime factor
-    count_a = Counter(factors_a)
-    count_b = Counter(factors_b)
-
     gcd = 1
     for factor in common_factors:
-        gcd *= factor ** min(count_a[factor], count_b[factor])
+        gcd *= factor ** min(factors_a.count(factor), factors_b.count(factor))
     
     return gcd
