+++ Math/Numerical_Methods/Constants/Pi_Algorithms/S_Ramanujan_algo.py
@@ -46,24 +46,20 @@ def calculate_pi_ramanujan(num_decimal_places: int = 50, num_terms: int = 10) ->
     # Calculate constant: (2√2)/9801
     constant = (Decimal(2) * Decimal(2).sqrt()) / Decimal(9801)
 
-    term_multiplier = Decimal(1)
-    # 396^4 = 24591257856
-    c396_4 = Decimal(24591257856)
-
     # Apply Ramanujan's series
     for k in range(num_terms):
+        # Calculate numerator: (4k)! × (1103 + 26390k)
+        numerator_factorial = factorial_decimal(4 * k)
         numerator_expression = Decimal(1103 + 26390 * k)
+        numerator = numerator_factorial * numerator_expression
 
-        # Add to total sum
-        term = term_multiplier * numerator_expression
+        # Calculate denominator: (k!)⁴ × 396^(4k)
+        denominator_factorial = (factorial_decimal(k)) ** 4
+        denominator_power = Decimal(396) ** (4 * k)
+        denominator = denominator_factorial * denominator_power
+
+        term = numerator / denominator
         total_sum += term
-
-        # Calculate next term_multiplier: T_{k} = (4k)! / ((k!)^4 * 396^{4k})
-        # The ratio T_{k+1} / T_k = (4k+4)(4k+3)(4k+2)(4k+1) / ((k+1)^4 * 396^4)
-        next_k = k + 1
-        num = Decimal((4*k + 4) * (4*k + 3) * (4*k + 2) * (4*k + 1))
-        den = Decimal(next_k**4) * c396_4
-        term_multiplier *= num / den
 
     # Calculate 1/π
     one_over_pi = constant * total_sum
