+++ Tests/test_factor_theorem.py
@@ -1,5 +1,6 @@
 import os
 import sys
+import pytest
 import math
 
 # Add root directory to path
